Extracted EditAttribute into its own file
